refactor: convert `<NuxtPicture>` to SFC
